[CPU] Refactoed directory with emitters
