+++ src/plugins/intel_cpu/src/emitters/snippets/x64/jit_container_emitter.cpp
@@ -1,0 +1,77 @@
+// Copyright (C) 2020-2023 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+//
+
+#include "jit_container_emitter.hpp"
+
+
+using namespace Xbyak;
+using namespace dnnl::impl;
+using namespace dnnl::impl::cpu::x64;
+
+namespace ov {
+namespace intel_cpu {
+
+jit_container_emitter::jit_container_emitter(dnnl::impl::cpu::x64::jit_generator* h, dnnl::impl::cpu::x64::cpu_isa_t isa) : jit_emitter(h, isa) {
+    in_out_type_ = emitter_in_out_map::gpr_to_gpr;
+}
+
+void jit_container_emitter::map_abstract_registers(mapping_info& gpr_map_pool, mapping_info& vec_map_pool,
+                                                   snippets::lowered::LinearIR::container& expressions) const {
+    if (expressions.empty())
+        OPENVINO_THROW("Cannot map registers when there is no allocated_emitters provided");
+
+    auto map_regs = [](const std::vector<size_t>& abstract_regs, mapping_info& mapping) {
+        auto& abstract_to_physical = mapping.first;
+        auto& regs_pool = mapping.second;
+        std::vector<size_t> physical_regs(abstract_regs.size());
+        for (size_t i = 0; i < abstract_regs.size(); i++) {
+            const auto abstract = abstract_regs[i];
+            auto& physical = physical_regs[i];
+            if (abstract_to_physical.count(abstract) == 0) {
+                if (regs_pool.empty())
+                    OPENVINO_THROW("Cannot map registers for jit_container_emitter: not enough regs in the pool");
+                physical = regs_pool.back();
+                regs_pool.pop_back();
+                abstract_to_physical[abstract] = physical;
+            } else {
+                physical = abstract_to_physical[abstract];
+            }
+        }
+        return physical_regs;
+    };
+
+    for (const auto& expression : expressions) {
+        const auto& emitter = expression->get_emitter();
+        std::vector<size_t> in_physical_regs, out_physical_regs;
+        std::vector<size_t> in_abstract_regs, out_abstract_regs;
+        std::tie(in_abstract_regs, out_abstract_regs) = expression->get_reg_info();
+        switch (std::dynamic_pointer_cast<jit_emitter>(emitter)->get_in_out_type()) {
+            case gpr_to_gpr:
+                in_physical_regs = map_regs(in_abstract_regs, gpr_map_pool);
+                out_physical_regs = map_regs(out_abstract_regs, gpr_map_pool);
+                break;
+            case gpr_to_vec:
+                in_physical_regs = map_regs(in_abstract_regs, gpr_map_pool);
+                out_physical_regs = map_regs(out_abstract_regs, vec_map_pool);
+                break;
+            case vec_to_gpr:
+                in_physical_regs = map_regs(in_abstract_regs, vec_map_pool);
+                out_physical_regs = map_regs(out_abstract_regs, gpr_map_pool);
+                break;
+            case vec_to_vec:
+                in_physical_regs = map_regs(in_abstract_regs, vec_map_pool);
+                out_physical_regs = map_regs(out_abstract_regs, vec_map_pool);
+                break;
+            default:
+                OPENVINO_THROW("Unsupported type of jit emitter!");
+        }
+        expression->set_reg_info({in_physical_regs, out_physical_regs});
+        if (auto container = std::dynamic_pointer_cast<jit_container_emitter>(expression->get_emitter()))
+            container->map_abstract_registers(gpr_map_pool, vec_map_pool, expressions);
+    }
+}
+
+
+}   // namespace intel_cpu
+}   // namespace ov
