+++ src/plugins/intel_cpu/src/emitters/snippets/x64/jit_perf_count_chrono_emitters.hpp
@@ -5,10 +5,10 @@
 
 #pragma once
 
-#include "jit_emitter.hpp"
-#include "cpu/x64/jit_generator.hpp"
+#include "emitters/common/x64/jit_emitter.hpp"
 
 #include "snippets/op/perf_count.hpp"
+
 
 namespace ov {
 namespace intel_cpu {
@@ -16,24 +16,26 @@ namespace intel_cpu {
 class jit_perf_count_chrono_start_emitter : public jit_emitter {
 public:
     jit_perf_count_chrono_start_emitter(dnnl::impl::cpu::x64::jit_generator *host, dnnl::impl::cpu::x64::cpu_isa_t host_isa,
-                            const std::shared_ptr<ov::Node>& n);
+                                        const std::shared_ptr<ov::Node>& n);
     size_t get_inputs_num() const override;
 
 private:
     void emit_impl(const std::vector<size_t> &in_idxs, const std::vector<size_t> &out_idxs) const override;
     static void set_start_time(snippets::op::PerfCountBegin* start_node);
+
     std::shared_ptr<snippets::op::PerfCountBegin> m_start_node = nullptr;
 };
 
 class jit_perf_count_chrono_end_emitter : public jit_emitter {
 public:
     jit_perf_count_chrono_end_emitter(dnnl::impl::cpu::x64::jit_generator *host, dnnl::impl::cpu::x64::cpu_isa_t host_isa,
-                            const std::shared_ptr<ov::Node>& n);
+                                      const std::shared_ptr<ov::Node>& n);
     size_t get_inputs_num() const override;
 
 private:
     void emit_impl(const std::vector<size_t> &in_idxs, const std::vector<size_t> &out_idxs) const override;
     static void set_accumulated_time(snippets::op::PerfCountEnd* end_node);
+
     std::shared_ptr<snippets::op::PerfCountEnd> m_end_node = nullptr;
 };
