+++ src/plugins/intel_cpu/src/emitters/snippets/x64/jit_perf_count_rdtsc_emitters.cpp
@@ -3,9 +3,7 @@
 //
 #ifdef SNIPPETS_DEBUG_CAPS
 
-#include "jit_emitter.hpp"
 #include "jit_perf_count_rdtsc_emitters.hpp"
-#include "cpu/x64/jit_generator.hpp"
 
 using namespace dnnl::impl;
 using namespace dnnl::impl::utils;
