+++ src/plugins/intel_cpu/src/nodes/interaction.cpp
@@ -8,8 +8,8 @@
 #include "cpu/x64/cpu_isa_traits.hpp"
 #include "cpu/x64/jit_generator.hpp"
 #include "dnnl_extension_utils.h"
-#include "emitters/x64/jit_dnnl_emitters.hpp"
-#include "emitters/x64/jit_load_store_emitters.hpp"
+#include "emitters/common/x64/jit_dnnl_emitters.hpp"
+#include "emitters/common/x64/jit_load_store_emitters.hpp"
 #include "interaction.h"
 #include "memory_desc/cpu_memory_desc_utils.h"
 #include "memory_desc/dnnl_blocked_memory_desc.h"
