+++ src/plugins/intel_cpu/src/nodes/interpolate.cpp
@@ -11,8 +11,8 @@
 #include "cpu/x64/jit_uni_eltwise.hpp"
 #include "dnnl_extension_utils.h"
 #include "eltwise.h"
-#include "emitters/x64/jit_bf16_emitters.hpp"
-#include "emitters/x64/jit_load_store_emitters.hpp"
+#include "emitters/common/x64/jit_bf16_emitters.hpp"
+#include "emitters/common/x64/jit_load_store_emitters.hpp"
 #include "fake_quantize.h"
 #include "onednn/dnnl.h"
 #include "openvino/core/parallel.hpp"
