+++ src/plugins/intel_cpu/src/nodes/mvn.cpp
@@ -14,8 +14,8 @@
 #include "dnnl_extension_utils.h"
 #include "utils/bfloat16.hpp"
 #include "openvino/core/parallel.hpp"
-#include "emitters/x64/jit_load_store_emitters.hpp"
-#include "emitters/x64/jit_bf16_emitters.hpp"
+#include "emitters/common/x64/jit_load_store_emitters.hpp"
+#include "emitters/common/x64/jit_bf16_emitters.hpp"
 
 #include "cpu/x64/jit_generator.hpp"
 #include "cpu/x64/jit_uni_eltwise.hpp"
