+++ src/plugins/intel_cpu/src/nodes/reduce.cpp
@@ -12,7 +12,7 @@
 #include "onednn/dnnl.h"
 #include "dnnl_extension_utils.h"
 #include "utils/bfloat16.hpp"
-#include "emitters/x64/jit_bf16_emitters.hpp"
+#include "emitters/common/x64/jit_bf16_emitters.hpp"
 #include "openvino/core/parallel.hpp"
 #include <algorithm>
