+++ src/plugins/intel_cpu/src/nodes/subgraph.h
@@ -6,11 +6,10 @@
 
 #include <ie_common.h>
 
-#include "onednn/dnnl.h"
-#include "cpu/x64/jit_generator.hpp"
-#include "emitters/x64/jit_snippets_emitters.hpp"
+#include <onednn/dnnl.h>
+#include "emitters/snippets/x64/jit_kernel_emitter.hpp"
 
-#include <node.h>
+#include "node.h"
 #include "snippets/op/subgraph.hpp"
 
 #include <array>
